[Chore #163023304] Rename src folder to server
