+++ test/contollers/meetup.test.js
@@ -1,6 +1,6 @@
 import chai, { assert } from 'chai';
 import app from '../../app';
-import statusCodes from '../../src/helpers/status';
+import statusCodes from '../../server/helpers/status';
 
 chai.use(require('chai-http'));
